weChat API

+++ src/main/java/com/isfinal/moudle/user/mapper/UserMapper.java
@@ -3,6 +3,7 @@ package com.isfinal.moudle.user.mapper;
 import com.isfinal.base.BaseMapper;
 import com.isfinal.moudle.user.model.User;
 
-public interface UserMapper extends BaseMapper<User,Integer>{
+public interface UserMapper extends BaseMapper<User,Long>{
+
 
 }
